+++ source/algorithms/class_infinite_DMRG.h
@@ -12,11 +12,11 @@
  *
  * \brief Class that runs the infinite DMRG algorithm.
  *
- * # infinite DMRG class
- * \param class_hdf5_file An hdf5 class object that handles the output file.
- * \param class_hdf5_table_buffer A buffer for table entries that goes into the output file
- * \param class_superblock A class that stores current MPS and environments at each iteration.
- * \param class_measurement A class that extracts, or measures, quantities from the superblock.
+ * # Infinite DMRG class
+ * \param shared_ptr<class_hdf5_file> An hdf5 class object that handles the output file.
+ * \param shared_ptr<class_hdf5_table_buffer> (optional) A buffer for table entries that goes into the output file
+ * \param shared_ptr<class_superblock> (optional) A class that stores current MPS and environments at each iteration.
+ * \param shared_ptr<class_measurement> (optional) A class that extracts, or measures, quantities from the superblock.
  */
 
 class class_infinite_DMRG : public class_base {
@@ -25,7 +25,7 @@ public:
     using class_base::class_base;
     explicit class_infinite_DMRG(std::shared_ptr<class_hdf5_file> hdf5_);
     long chi_max    = settings::idmrg::chi_max;
-    long max_length = settings::idmrg::max_length;
+    int  max_length = settings::idmrg::max_length;
     int  print_freq = settings::idmrg::print_freq;
     int  iteration  = 0;
